Add tests for the pwn proc context
